% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distributions.R
\name{add_distributions}
\alias{add_distributions}
\title{Add together two epiwave_distribution_massfun objects}
\usage{
add_distributions(delay1, delay2)
}
\arguments{
\item{delay1}{first epiwave dist object to combine}

\item{delay2}{second epiwave dist object to combine}
}
\value{
epiwave_distribution_massfun
}
\description{
Add together two epiwave_distribution_massfun objects
}
