% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epiwave_timeseries.R
\name{create_epiwave_greta_timeseries}
\alias{create_epiwave_greta_timeseries}
\title{Expand distribution into long tibble}
\usage{
create_epiwave_greta_timeseries(dates, jurisdictions, car, chr_prior)
}
\arguments{
\item{dates}{infection dates sequence}

\item{jurisdictions}{jurisdiction names}

\item{car}{x}

\item{chr_prior}{x}
}
\value{
list
}
\description{
The epiwave model functions expect data in a long format,
which is structured to have a value for every unique date and jurisdiction
pair. This function create a tibble of this structure out of a single
distribution that should be replicated in each cell.
}
