% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_distribution.R
\name{data_to_distribution}
\alias{data_to_distribution}
\title{Create distribution object from data}
\usage{
data_to_distribution(
  data_date_1,
  data_date_2,
  min_delay_days = NULL,
  max_delay_days = NULL,
  min_delay_quantile = 0,
  max_delay_quantile = 0.99
)
}
\arguments{
\item{data_date_1}{data column with first date to calculate delay from (earlier date)}

\item{data_date_2}{data column with second date to calculate delay from (later date)}

\item{min_delay_days}{optional specification for minimum delay}

\item{max_delay_days}{optional specification for maxiumum delay}

\item{min_delay_quantile}{optional specification for quantile to select
minimum delay IF min_delay_days hasn't been set. Defaults to 0.}

\item{max_delay_quantile}{optional specification for quantile to select
maxiumum delay IF max_delay_days hasn't been set. Defaults to 0.99.}
}
\value{
epiwave_distribution object
}
\description{
Create distribution object from data
}
