% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epiwave_massfun.R
\name{plot.epiwave_massfun}
\alias{plot.epiwave_massfun}
\title{Default plot function for epiwave_massfun object}
\usage{
\method{plot}{epiwave_massfun}(x, y, ...)
}
\arguments{
\item{x}{x value}

\item{y}{y value}

\item{...}{additional arguments}
}
\value{
plot
}
\description{
Default plot function for epiwave_massfun object
}
