% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epiwave_massfun.R
\name{parametric_dist_to_distribution}
\alias{parametric_dist_to_distribution}
\title{Create distribution object from parametric distribution values}
\usage{
parametric_dist_to_distribution(dist, min_delay = NULL, max_delay = NULL)
}
\arguments{
\item{dist}{distributional package object}

\item{min_delay}{optional specification for minimum delay}

\item{max_delay}{optional specification for maxiumum delay}
}
\value{
epiwave_distribution object
}
\description{
Create distribution object from parametric distribution values
}
